clean up sidebar logic
